dashboard/requirements.txt
flask
google-cloud-firestore
requests
